Added Jump Noise
Added code for making a sound every time the player jumps, as well as modifying how the sounds are stored to allow them to be used by the game.
Also added a .gitignore, since I was having trouble with the online phaser source on my end and ended up downloading it to a local file.