Changed bullet variable names
Made the shots fired by the bosses have different names, and the ones fired by the player have the same name, to ensure that the bosses could fire their unique weapons and the player's shots stayed consistent. Also, fixed a typo in the asset name for the boss' bullets in level 4.